+++ HISTORY.rst
@@ -12,6 +12,9 @@ Unreleased
 
 **Build**
 
+* ALCN-1087: Disable Camptocamp platform specifics by default
+  plus add an opt-in with a new environment variable C2C_PLATFORM
+
 **Documentation**